[Karate-Tools] Sonar Issues Fix
